refactor ic-boundary
