+++ rs/boundary_node/ic_boundary/src/main.rs
@@ -5,19 +5,17 @@ use tikv_jemallocator::Jemalloc;
 use crate::cli::Cli;
 
 mod bouncer;
-mod cache;
 mod check;
 mod cli;
 mod core;
 mod dns;
+mod errors;
 mod firewall;
-mod geoip;
 mod http;
 mod log;
 mod metrics;
 mod persist;
 mod rate_limiting;
-mod retry;
 mod routes;
 mod salt_fetcher;
 mod snapshot;
