+++ rs/boundary_node/ic_boundary/src/rate_limiting/generic.rs
@@ -44,8 +44,9 @@ use super::{
 
 use crate::{
     core::Run,
+    errors::{ErrorCause, RateLimitCause},
     persist::RouteSubnet,
-    routes::{ErrorCause, RateLimitCause, RequestContext, RequestType},
+    routes::{RequestContext, RequestType},
     snapshot::RegistrySnapshot,
 };
